+++ src/SpriteSheet.cpp
@@ -6,7 +6,7 @@
  * SpriteSheet - Given a filename, a frame height and frame width, load a
  * sprite sheet into memory with the assumed given frame size.
  */
-SpriteSheet::SpriteSheet(std::string filename, uint width, uint height) {
+SpriteSheet::SpriteSheet(std::string filename, uint16_t width, uint16_t height) {
 	this->sheet  = IMG_Load(filename.c_str());
 
 	this->width  = width;
@@ -16,14 +16,14 @@ SpriteSheet::SpriteSheet(std::string filename, uint width, uint height) {
 /* ------------------------------------------------------------------------------
  * getWidth - Return the width of the frame size.
  */
-uint SpriteSheet::getWidth() {
+uint16_t SpriteSheet::getWidth() {
 	return this->width;
 }
 
 /* ------------------------------------------------------------------------------
  * getHeight - Return the height of the frame size.
  */
-uint SpriteSheet::getHeight() {
+uint16_t SpriteSheet::getHeight() {
 	return this->height;
 }
 
@@ -32,7 +32,7 @@ uint SpriteSheet::getHeight() {
  * it to the provided SDL_Rect on the screen.  Blocks until the screen is
  * unlocked.
  */
-void SpriteSheet::blitFrame(uint frameX, uint frameY, SDL_Rect& where) const {
+void SpriteSheet::blitFrame(uint16_t frameX, uint16_t frameY, SDL_Rect& where) const {
 
 	// Source Rect
 	SDL_Rect srcRect;
